Fixed a major bug in arduino serial escaping
 - was not casting result of (input ^ BEAN_ESCAPE_CHAR) to a uint8_t
 - HardwareSerial::write() does very funny stuff when this is not casted
 - results were buggy whenever sending chars that needed to be escaped

+++ hardware/bean/cores/bean/BeanSerialTransport.cpp
@@ -283,8 +283,8 @@ inline void BeanSerialTransport::insert_escaped_char(uint8_t input){
     case BEAN_EOF:  // fallthrough
     case BEAN_ESCAPE:
       HardwareSerial::write(BEAN_ESCAPE);
-      HardwareSerial::write(input ^ BEAN_ESCAPE_XOR);
-    break;
+      HardwareSerial::write(uint8_t(input ^ BEAN_ESCAPE_XOR));
+      break;
     default:
       HardwareSerial::write(input);
   }
